- Eliminated more truncation and conversion between types warnings under MSVC.
- Moved some code to use qsqrt() rather than sqrt() where appropriate.

Note:

MSVC does not like float value's being expressed as ".0" or "0." and always likes you
to use a trailing "f" if you are allocating a value to a the float type.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@9407 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csgeom/quaterni.h
@@ -79,7 +79,7 @@ public:
   void PrepRotation (float angle, csVector3 vec)
   {
     double theSin = sin (angle / 2);
-    Init (cos (angle / 2), vec.x*theSin, vec.y*theSin, vec.z*theSin);
+    Init ((float) cos (angle / 2), vec.x * theSin, vec.y * theSin, vec.z * theSin);
   }
 
   /// rotated = q * vec * qConj.
